+++ src/utils/pug-utils.ts
@@ -1,6 +1,8 @@
 // Partial copy of https://github.com/prettier/plugin-pug/blob/main/src/utils/common.ts
 
 import type { AttributeToken, TagToken, Token } from 'pug-lexer';
+import * as lex from 'pug-lexer';
+import { findIndexFrom } from './index';
 
 /**
  * Returns the previous tag token if there was one.
@@ -64,6 +66,18 @@ export function previousTypeAttributeToken(tokens: ReadonlyArray<Token>, index: 
     }
   }
   return;
+}
+
+export function getAttributeTokens(tag: TagToken, tokens: ReadonlyArray<lex.Token>): AttributeToken[] {
+  const tagIndex: number = tokens.indexOf(tag);
+  const startAttributesIndex: number = findIndexFrom(tokens, ({ type }) => type === 'start-attributes', tagIndex);
+  const endAttributesIndex: number = findIndexFrom(
+    tokens,
+    ({ type }) => type === 'end-attributes',
+    startAttributesIndex
+  );
+
+  return tokens.slice(startAttributesIndex + 1, endAttributesIndex) as AttributeToken[];
 }
 
 /**
